sql: fix LATERAL-related panic in try_simplify_quantified_comparison
The try_simplify_quantified_comparison function, which optimizes a SQL
expression pre-decorrelation, needs to walk a SQL expression while
keeping track of the outer relation type. It was not properly updated
for the addition of LATERAL joins, which are an additional expression
variant that introduce a new inner scope.

If we start writing more SQL optimizations, we'll want to figure out how
to package up this logic so it only needs to be written in one place.
But for now, hardcoding it here is the simple fix.

Fix #3924.