coord,dataflow: un-special-case local inputs
The only difference between a local source and an external source is
that data comes from a local Timely unordered input rather than Kafka or
a file. Yet the coordinator and dataflow layers had a separate API for
describing how to build a local input. This complicated the code
unnecessarily, so remove it. Local sources are now built and exported as
an index via DataflowDescs, just like for external sources.

Supersedes #3888.