deps: bump to redox_users v0.3.5
We don't support RedoxOS, but the redox_users crate gets pulled in as
part of the `dirs` crate, which is a cross-platform abstraction layer
over finding user's home directories and such. Bump to v0.3.5 to remove
an old version of the base64 crate from our dependency set.

We're inching ever closer to having no duplicate dependencies in our
tree.