Use monoids for monotonic min/max (#4122)
This PR uses differential dataflow's support for monoids (in place of integer counts) to accumulate min and max operators with monotonic (non-removing) inputs. This is appropriate for mins and maxes in a way that doesn't easily generalize to top k (at least, for non-trivial k), so trying it out here first.

There may be some iteration as errors are surfaced. In particular, this implementation will perform less type-based error checking than the prior implementation, which would panic on type errors (this implementation just goes with the smallest non-Datum::Null value, which could in principle be of different types if we have errors elsewhere in the system).