Fix hanging sqllogictest runner (#4167)
Since I added the mz_catalog.mz_columns system table in #4085, the sqllogictest runner
has been hanging locally. Adding the mz_columns table meant that we were broadcasting
a new insert for every builtin column in a Materialize instance.

Because the sqllogictest runner tried to instantiate a Coordinator before it
launched its dataflow workers, the Coordinator was sending tons of messages into the
comm channel without anything to read them back out. This issue was fixed for actual
Materialize instances with #3284 by simply launching the dataflow workers first. However,
this fix was not applied to the sqllogictest runner at that time.

This PR updates the sqllogictest runner to launch its dataflow workers before it attempts
to instantiate the Coordinator, mirroring the fix in #3284.