prof: simplify installation of malloc_conf
Export `malloc_conf` as a `&[u8]` rather than a `Option<&c_char>`. This
seems a good bit less magical.

Based on my read of the Rust reference's "Type Layout" chapter, pointers
and references have the same layout, and obviously a u8/i8s have the
same layout, so a symbol of type `&[u8]` should be exactly equivalent to
a symbol of type `*c_char`. The use of an option type in the old
approach actually strikes me as kind of sketchy; not sure what the
representational guarantees of `Option` are.

I've manually verified that this new approach works on Linux.