dataflow: change file name format for persisted files.
Previously, persisted files were named [start, end] where `start` and `end`
denoted the range of offsets stored in this file as 1-indexed Kafka offsets.

This was downright confusing because Kafka offsets are actually 0-indexed, and
storing them in a 1-indexed format is a (hopefully temporary) Materialize-ism.

This commit tries to preserve as much existing behavior as possible while making
the generated file names for persisted files more understandable to the end user.
Now, file names have [start, end) where both are 0-indexed offsets, and `end` is the
next offset after the last one stored in the file.