dataflow: prefer hashing positions when distributing uniformly across workers
Closes #4197

Previously, when we wanted to distribute records across workers uniformly
randomly, we distributed them based on the hash of their values. The reasoning
behind this was that in many potential sources, the distribution of keys might
be sufficiently skewed (or keys may not even exist) which makes keys a poor
choice uniformly distributing data (except in cases where we need records
with the same key to go to the same worker).

Unfortunately, all of those same arguments apply for values as well. This commit
changes the behavior to distribute records based on the hash of the position if
the position exists and otherwise fall back to a hash on the value. The
distribution of positions is much better known to us than the distribution of
potential values, and its more likely to be amenable to what we are trying to
accomplish.