doc/user: add CPU requirement to the Docker configuration guide
It turns out CPU count is important for running demos, too, so note that
in the "Using Docker" doc in addition to the BI demo doc. Also adjust
the wording in the BI demo doc to be a bit more precise about what
precisely needs its resources adjusted (i.e., Docker for Mac).