bin/pycheck: fix bugs
There is no `--ci` argument to bin/pyactivate. CI mode is activated by
setting the "CI" environment variable to a truthy value. This wasn't
properly failing the script because the `try` command was executed in a
subshell. Fix that by passing the sysroot to `pytest` as a command line
argument, rather than changing the CWD.