sql: support array types
Split from #4305. Touches #1248.

This commit adds enough support for PostgreSQL-style array types to the
SQL layer that they can be constructed (in experimental mode) and sent
over the wire in either the text or binary format.

The goal here is to unlock pg_catalog views that require arrays, not to
make these arrays generally usable for end users. Users should in
general use lists, not arrays.