deps: always compile adler + rustc-demangle with optimizations
adler in particular is critical to the speed of backtrace generation.
If compiled without optimizations, generating a backtrace in a debug
build takes nearly a minute. With optimizations, generating a backtrace
takes only a few seconds.

rustc-demangle is included for good measure, so that the rule is that
any transitive dependency of the backtrace crate is compiled with
optimizations, even in debug builds.