coord,dataflow: tie persistence data to catalog versions (cluster_id)
Closes #4029

Previously, we had a ABA type of problem in source persistence where a user
could take some previously persisted data, and swap out the corresponding catalog,
at which point its possible that source global ids used in the persistent data
no longer correspond to those in the catalog.

This change makes it so that all persistent data remembers what catalog
version it was created from, and refuses to ingest persistent data that does not
belong to the same catalog.