coord: don't tie all views to presence of logging
Many, but not all, built-in views depend on the availability of the
built-in logging sources. We previously disabled *all* views when
logging was disabled, since you can't install a view on a logging source
if the logging source doesn't exist, but now we've got a bunch of
pg_catalog views that don't depend on the logging sources. Get a bit
smarter and learn to only disable the views that depend on logging
sources when logging is disabled.