sql: handle subqueries in sql_op!-based functions
The sql_op! macro permits defining a function whose implementation is a
SQL fragment that is executed in the context of the containing query.
This patch teaches that macro to handle SQL fragments that contain
subqueries.

This new behavior unlocks some awesome expressivity, as subqueries can
be correlated, and thus the implementation of a function can elegantly
introduce a join, which will be important for some pgcompat functions.
Unfortunately the implementation is a bit complicated, as it requires
correcting column indexes for any correlated column references.