Improve Debezium duplicate-detection diagnostics
With this change we only emit `warn` logs when Debezium guarantees are violated.

This preserves an info log when we see a surprising but legal message -- something behind
the Debezium event frontier.

This increases the state size for the track-full case, but a future PR will allow us to
only temporarily enable full-id tracking, permanently flattening memory usage.