pgwire: extract connection helper object
Extract a dedicated FramedConn object that manages the sending and
receiving of pgwire messages. Previously the protocol module had to
build this up out of various Tokio and pgwire parts; this path
centralizes that responsibility in the codec module.

Most importantly, this scopes the borrows on the `send`/`recv` family of
methods to just the `conn` field of `StateMachine`, which is important
for the next commit.