coord: introduce a client layer
The coordinator's correctness relies on the fact that it runs on its own
Tokio task (i.e., a thread equivalent). Interacting with the coordinator
therefore requires sending commands over a channel, and then waiting
upon a response channel.

This patch introduces a coord::Client abstraction layer that manages the
communication with the coordinator task. This greatly simplifies all
code that interacts with the coordinator--pgwire, the HTTP server, and
the sqllogictest runner--since complicated channel send and receives are
replaced with simple method calls.

Also upgrade to a more recent version of node-pg. Apparently the version
we were using had a bug that was tickled by this refactor.