sql: ditch "global ID" nomenclature for new tables
The term "global ID" doesn't make much sense to end users. Global to
what? (Internally, it distinguishes between IDs in an expression that
are local to that expression and those that are not.) Just use the
phrasing "ID" instead, which is consistent with what we do for databases
and schemas.