sql: eagerly coerce arrays and lists
In PostgreSQL, a bare string literal like

    '42'

will have type "unknown". If that string literal is used in an
expression like

    '42' + 1

PostgreSQL will "coerce" it to an integer, and the expression is
therefore well-typed.

("unknown" types are confusing, though, so we use something slightly
different in Materialize: we explicitly represent expressions whose type
is undetermined via the CoercibleScalarExpression type.)

A similar situation occurs with bare ARRAY expressions, like:

    ARRAY['1', '2']

Unlike string literals, in PostgreSQL, ARRAY expressions are eagerly
coerced, and so this expression is immediately determined to have type
"text[]". This determination is made by forcibly coercing the elements
with, so

    ARRAY[1, '2']

will be assigned type "int[]", since the usual rules for homogeneous
coercion apply to the elements. This poses a bit of a problem for
empty arrays like

    ARRAY[]

since there are no elements that can determine the element type. That
expression alone yields a "cannot determine type of empty array" error.

So PostgreSQL special cases direct casts of ARRAY expressions, as in

    ARRAY[]::int[]

so that you can at least construct empty arrays of a desired type.  More
complicated cases, like `function_accepting_int_array(ARRAY[])`, are not
handled, and result in the same "cannot determine type of empty array"
error.

Previously, Materialize was a bit too smart, and typed empty ARRAY
expressions as "unknown", precisely so that something like
`function_accepting_int_array(ARRAY[])` would be accepted.  These extra
smarts interact poorly with polymorphic function arguments (#4582),
however, as resolving polymorphic function arguments requires knowing
the concrete array/list types involved.

So, just do it exactly like PostgreSQL does it. ARRAY and LIST
expressions are now eagerly coerced, with one special case for when an
ARRAY/LIST expression is directly cast to a different type.  While this
is a bit less expressive, it results in only minor regressions in our
test suite, and support for polymorphic function arguments is well worth
it.