materialized: tweak startup logs
Rearrange startup so that the splash screens and "listening on..."
message are printed at the end of startup, where they are more likely to
be visible. Previously they'd get lost above a bunch of info logs, if
logging to stderr.

Also include more information about the OS/hardware on which Materialize
is running.

Fix #4603.