repr: fix undefined behavior in RowArena::push_row
RowArena::push_row resulted in undefined behavior, since the
reference to the returned Row could be invalidated if
RowArena::owned_rows reallocated. In practice, we didn't trigger the UB,
because every call to push_row had the following form:

    arena.push_row(row).unpack_first()

That is, we immediately unpacked the returned row, and only held onto a
reference to the first datum in the row. This *is* ok, because even if
owned_rows reallocates and the row moves, the data in the row will not.
Adjust the API of RowArena to only support this pattern via

    fn push_unary_row(&'a self, row: Row) -> Datum<'a>

which is actually a safe API that cannot be misused.

h/t @umanwizard for spotting the UB.