sql: represent operators as text in AST
Rather than representing operators via two enums, UnaryOperator and
BinaryOperator, represent them as a string that stores the literal
operator symbol. This brings our implementation more in line with
PostgreSQL, which permits users to create arbitrary operators out of a
set of allowable operator symbols.

Also rework the function selection API a bit to share more code.
Operators and functions are now basically indistinguishable from the
perspective of the sql::plan::func module.

The goal here is to make room for PostgreSQL's qualified operator
syntax:

    SELECT 1 OPERATOR(pg_catalog.+) 2

though actually implementing this syntax is left to a forthcoming
commit.