sql: make TAIL a normal rows-sending statement
TAIL has transitioned to be send rows similar to a SELECT. It has
grown two new columns: timestamp and diff. Peek and Tail still operate
differently, but they implement a common interface that can be used by
the row send operation.

Additionally, COPY TO has been taught how to send TAIL operations such
that TAIL can be accessed similarly to how it was before, but now with
greater ability to parse the output columns.

Closes #4540
Closes #3091