sql: allow "assignment" casts when INSERT types don't match
PostgreSQL will automatically cast values in an INSERT statement that
don't match the type of the table into which they're being inserted.
The casts allowed in this context are called "assignment" casts, which
are a superset of "implicit" casts and a subset of "explicit" casts.

This is particularly important for the common case of inserting numbers
into float columns, as in:

    CREATE TABLE t (a float);
    INSERT INTO t VALUES (1234.0);

Previously, Materialize would reject the INSERT, because "1234.0" had
type "decimal(38, 1)" while the target column had type "float". With
this patch, an assignment cast from float to decimal is automatically
injected, and the query succeeds.

This patch also reclassifies many explicit casts as assignment casts,
based on what PostgreSQL 13 declares in the pg_cast table.

Fix #2852.